chore: finished migrating pnpm, fixed local build and prepared monorepo build pipeline
