+++ apps/api/deploy-api.yml
@@ -11,41 +11,85 @@ parameters:
   type: string
 - name: vmImageName
   type: string
-- name: functionAppNamePri
+- name: appSettingsJsonFileRelativePathPri
+  displayName: 'AppSettings Json File Relative Path for Primary instance'
   type: string
 
 jobs:
 - job: Infrastructure
   displayName: Infrastructure Setup
-  condition: eq(stageDependencies.Build.BuildJob.outputs['BuildJob.HAS_BACKEND_CHANGES'], 'true')
+  condition: eq(stageDependencies.Build.Build.outputs['BuildJob.HAS_BACKEND_CHANGES'], 'true')
   pool:
     vmImage: ${{parameters.vmImageName}}
   steps:
   - task: AzureResourceManagerTemplateDeployment@3
+    name: deployInfrastructure
+    displayName: 'Deploy API Infrastructure'
     inputs:
       connectedServiceName: ${{parameters.ServiceConnectionName}}
-      deploymentName: $(Build.BuildNumber)
+      deploymentName: $(Build.BuildNumber)-api
       location: ${{parameters.deploymentDefaultLocation}}
       resourceGroupName: ${{parameters.resourceGroupName}}
       csmFile: apps/api/iac/main.bicep
-      overrideParameters: >
-                -environment ${{parameters.environmentNameBicep}}
+      csmParametersFile: apps/api/iac/${{parameters.environmentNameBicep}}.bicepparam
+      deploymentOutputs: 'armOutputs'
+      # overrideParameters: >
+      #           -environment ${{parameters.environmentNameBicep}}
+  
+  - task: PowerShell@2
+    name: captureOutputs
+    displayName: 'Capture Infrastructure Outputs'
+    inputs:
+      targetType: 'inline'
+      script: |
+        Write-Host "Debug - armOutputs: $(armOutputs)"
+        $var = ConvertFrom-Json '$(armOutputs)'
+        $functionAppNamePri = $var.functionAppNamePri.value
+        Write-Host "Function App Name (Primary): $functionAppNamePri"
+        Write-Host "##vso[task.setvariable variable=functionAppNamePri;isOutput=true]$functionAppNamePri"
+
 - deployment: Application_Deployment_Pri
   displayName: Application Deployment (Primary)
-  condition: eq(stageDependencies.Build.BuildJob.outputs['BuildJob.HAS_BACKEND_CHANGES'], 'true')
+  condition: and(succeeded(), eq(stageDependencies.Build.Build.outputs['BuildJob.HAS_BACKEND_CHANGES'], 'true'))
+  dependsOn: Infrastructure
   environment: ${{parameters.environmentNameDevOps}}
   pool:
     vmImage: ${{parameters.vmImageName}}
+  variables:
+    functionAppNamePri: $[ dependencies.Infrastructure.outputs['captureOutputs.functionAppNamePri'] ]
   strategy:
     runOnce:
       deploy:
         steps:
+        - download: current
+          displayName: 'Artifact: Download API package'
+          artifact: api
         - task: AzureFunctionApp@1
-          displayName: 'Azure Functions App Deploy: ${{parameters.functionAppNamePri}}'
+          displayName: 'Azure Functions App Deploy: $(functionAppNamePri)'
           inputs:
             azureSubscription: ${{parameters.ServiceConnectionName}}
             appType: 'functionAppLinux'
-            appName: ${{parameters.functionAppNamePri}}
-            package: '$(Pipeline.Workspace)/drop/$(Build.BuildId).zip'
+            appName: $(functionAppNamePri)
+            package: '$(Pipeline.Workspace)/api/api-$(Build.BuildId).zip'
             runtimeStack: 'NODE|22'
             deploymentMethod: 'runFromPackage'
+
+- job: Application_Settings_Pri
+  displayName: Application Settings (Primary)
+  condition: and(succeeded(), eq(stageDependencies.Build.Build.outputs['BuildJob.HAS_BACKEND_CHANGES'], 'true'))
+  dependsOn: 
+    - Infrastructure
+    - Application_Deployment_Pri
+  pool:
+    vmImage: ${{parameters.vmImageName}}
+  variables:
+    functionAppNamePri: $[ dependencies.Infrastructure.outputs['captureOutputs.functionAppNamePri'] ]
+  steps:
+  - powershell: |
+      $output = Get-Content '$(Build.SourcesDirectory)/${{parameters.appSettingsJsonFileRelativePathPri}}'
+      Write-Host "##vso[task.setvariable variable=appSettingsPri]$($output)"
+  - task: AzureAppServiceSettings@1
+    inputs:
+      azureSubscription: ${{parameters.ServiceConnectionName}}
+      appName: $(functionAppNamePri)
+      appSettings: '$(appSettingsPri)'
