--- iac/app-service-plan/main.bicepparam
@@ -1,11 +1,0 @@
-using './main.bicep'
-
-param appServicePlanName = 'pri-001'
-param location = 'eastus2'
-param tags = {
-  environment: 'dev'
-  application: 'cel'
-}
-param sku = 'EP1'
-param operatingSystem = 'linux' // Options: linux, windows
-
