+++ iac/cosmos-mongodb/mongo-database.bicep
@@ -10,10 +10,10 @@ param maxThroughput int
 @description('Tags')
 param tags object
 
-resource mongoDatabaseAccount 'Microsoft.DocumentDB/databaseAccounts@2022-08-15' existing = {
+resource mongoDatabaseAccount 'Microsoft.DocumentDB/databaseAccounts@2025-05-01-preview' existing = {
   name: mongoDatabaseAccountName
 }
-resource mongoDatabase 'Microsoft.DocumentDB/databaseAccounts/mongodbDatabases@2022-08-15' = {
+resource mongoDatabase 'Microsoft.DocumentDB/databaseAccounts/mongodbDatabases@2025-05-01-preview' = {
   parent: mongoDatabaseAccount
   name: mongoDatabaseName
   tags: tags
