+++ iac/function-app/main.bicepparam
@@ -4,15 +4,17 @@ param appServicePlanName = 'asp-pri-001'
 param location = 'eastus2'
 param tags = {
   environment: 'dev'
-  application: 'cellix'
+  application: 'sharethrift'
 }
 param storageAccountName = 'cxadevstfunceastus2'
 param maxOldSpaceSizeMB = 3072
 param functionWorkerRuntime = 'node'
-param linuxFxVersion = 'NODE|20' 
+param linuxFxVersion = 'NODE|22' 
 param functionExtensionVersion = '~4' 
-param allowedOrigins = [] //Specify the frontend endpoints that can access the function app
-param keyVaultName = 'corp-dev-kv'
+param allowedOrigins = [
+  'https://sth-fde-ged3a8gxcvfxafaf.z02.azurefd.net'
+] //Specify the frontend endpoints that can access the function app
+param keyVaultName = 'sharethrift-keyvault'
 param functionAppInstanceName = 'pri'
-param applicationPrefix = 'cel'
+param applicationPrefix = 'sth'
 param env = 'dev'
