+++ iac/global/resource-naming-convention.bicep
@@ -9,3 +9,6 @@ var resourceTypes = loadJsonContent('./resource-types.json')
 output smallPrefix string = '${applicationPrefix}${environment}'  // 6 chars length
 output prefix string = '${applicationPrefix}-${environment}-'    // 8 chars length
 output resourceTypes object = resourceTypes
+
+// Microsoft Resource Naming Conventions
+// https://learn.microsoft.com/en-us/azure/cloud-adoption-framework/ready/azure-best-practices/resource-abbreviations
