+++ iac/static-website/front-door.bicep
