+++ iac/static-website/main.bicep
@@ -21,26 +21,8 @@ param storageAccountLocation string
 ])
 param storageAccountSku string
 
-@allowed([
-  'eastus2'
-])
-param cdnLocation string
-
-@description('CDN SKU Names')
-@allowed([
-  'Standard_Akamai'
-  'Standard_Microsoft'
-  'Standard_Verizon'
-  'Premium_Verizon'
-])
-param cdnSku string
-
-param googleAnalyticsSha256 string
-
 @maxLength(3)
 param instanceName string
-
-param customDomainName string
 
 @description('Array of allowed origins for CORS.')
 param corsAllowedOrigins array
@@ -52,8 +34,6 @@ param corsAllowedHeaders array
 param corsExposedHeaders array
 @description('maxAge for CORS')
 param corsMaxAgeInSeconds int
-@description('Array of cdn rules')
-param cdnRules array
 @description('Tags')
 param tags object
 
@@ -65,7 +45,6 @@ param tags object
 @description('Specifies whether data in the container may be accessed publicly and the level of access.')
 param publicAccessLevel string = 'None'
 @description('specify cdn profile to associate with the storage account')
-param cdnProfileName string
 
 // variables
 var uniqueId = uniqueString(resourceGroup().id)
@@ -98,22 +77,20 @@ module storageAccountModule './storage-account.bicep' = {
   }
 }
 
-// cdn
-module cdnModule './cdn.bicep' = {
-  name: '${applicationPrefix}${instanceName}cdnModule'
-  dependsOn: [storageAccountModule]
-  params: {
-    applicationPrefix: applicationPrefix
-    cdnProfileName: cdnProfileName
-    location: cdnLocation
-    cdnSku: cdnSku
-    cdnRules: cdnRules
-    cdnEndpointName: '${resourceNamingConvention.outputs.prefix}${resourceNamingConvention.outputs.resourceTypes.cdnEndpoint}-${instanceName}-${uniqueId}'
-    storageAccountPrimaryHostName: replace(replace(storageAccountModule.outputs.storageAccountPrimaryEndpointWeb, 'https://', ''), '/','')
-    storageAccountSecondaryHostName: replace(replace(storageAccountModule.outputs.storageAccountSecondaryEndpointWeb, 'https://', ''), '/','')
-    googleAnalyticsSha256: googleAnalyticsSha256
-    customDomainName: customDomainName
-    tags: tags
-  }
-}
-                                   
+// // cdn
+// module cdnModule './cdn.bicep' = {
+//   name: '${applicationPrefix}${instanceName}cdnModule'
+//   dependsOn: [storageAccountModule]
+//   params: {
+//     applicationPrefix: applicationPrefix
+//     cdnProfileName: cdnProfileName
+//     cdnRules: cdnRules
+//     cdnEndpointName: '${resourceNamingConvention.outputs.prefix}${resourceNamingConvention.outputs.resourceTypes.cdnEndpoint}-${instanceName}-${uniqueId}'
+//     storageAccountPrimaryHostName: replace(replace(storageAccountModule.outputs.storageAccountPrimaryEndpointWeb, 'https://', ''), '/','')
+//     storageAccountSecondaryHostName: replace(replace(storageAccountModule.outputs.storageAccountSecondaryEndpointWeb, 'https://', ''), '/','')
+//     customDomainName: customDomainName
+//     tags: tags
+//   }
+// }
+
+output storageAccountName string = storageAccountModule.outputs.storageAccountName
