--- iac/storage-account/main.bicepparam
@@ -1,11 +1,0 @@
-using './main.bicep'
-
-param env = 'dev'
-param storageAccountName = 'ahpstapp'
-param location = 'eastus'
-param skuName = 'Standard_LRS'
-param kind = 'StorageV2'
-param tags = {
-  environment: env
-  application: 'ahp'
-}
