+++ packages/sthrift/graphql/package.json
@@ -23,21 +23,25 @@
     },
     "dependencies": {
         "@apollo/server": "^4.11.3",
+        "@apollo/utils.withrequired": "^3.0.0",
         "@as-integrations/azure-functions": "^0.2.0",
-        "@azure/functions": "^4.6.0",
+        "@azure/functions": "^4.0.0",
         "@graphql-tools/json-file-loader": "^8.0.20",
         "@graphql-tools/load": "^8.1.2",
         "@graphql-tools/load-files": "^7.0.1",
+        "@graphql-tools/merge": "^9.1.1",
+        "@graphql-tools/schema": "^10.0.25",
+        "@opentelemetry/api": "^1.9.0",
+        "@sthrift/application-services": "workspace:*",
+        "@sthrift/domain": "workspace:*",
         "graphql": "^16.10.0",
         "graphql-middleware": "^6.1.35",
-        "graphql-scalars": "^1.24.2",
-        "@sthrift/application-services": "workspace:*",
-        "@sthrift/domain": "workspace:*"
+        "graphql-scalars": "^1.24.2"
     },
     "devDependencies": {
         "@cellix/typescript-config": "workspace:*",
         "@cellix/vitest-config": "workspace:*",
-        "typescript": "^5.8.3",
-        "rimraf": "^6.0.1"
+        "rimraf": "^6.0.1",
+        "typescript": "^5.8.3"
     }
 }
