+++ packages/sthrift/persistence/package.json
@@ -26,13 +26,14 @@
         "@cellix/domain-seedwork": "workspace:*",
         "@cellix/event-bus-seedwork-node": "workspace:*",
         "@cellix/mongoose-seedwork": "workspace:*",
+        "@sthrift/data-sources-mongoose-models": "workspace:*",
         "@sthrift/domain": "workspace:*",
-        "@sthrift/data-sources-mongoose-models": "workspace:*"
+        "bson": "^6.10.4"
     },
     "devDependencies": {
         "@cellix/typescript-config": "workspace:*",
         "@cellix/vitest-config": "workspace:*",
-        "typescript": "^5.8.3",
-        "rimraf": "^6.0.1"
+        "rimraf": "^6.0.1",
+        "typescript": "^5.8.3"
     }
 }
